backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
pydantic[email]>=1.8.2
numpy>=1.20.0
pandas>=1.3.0
rdkit>=2021.03.1
neo4j>=4.4.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
python-dotenv>=0.19.0

# Authentication dependencies
passlib>=1.7.4
bcrypt>=3.2.0
PyJWT>=2.1.0
email-validator>=1.1.3
